remove tally and range functions for consistent dplyr syntax
while the tally function is convenient, it deviates
from the standard split apply combine syntax that
dplyr used: namely, group-by one variable (or more)
and apply summary functions to this grouped dataframe.
While the n() function is peculiar in that it takes no
arguments, it does not fail due to NAs and so is the smallest
increment I can think of. Using another function like
"mean" would increase the  cognitive load on the
students too much due to the inclusion of NAs.

also removed range as a potential aggregating function
as it requires more complicated syntax to use this as part
of the split apply combine approach.